+++ server/src/main/java/replicant/server/transport/ReplicantSessionAuthorization.java
@@ -6,6 +6,12 @@ import org.jspecify.annotations.NonNull;
 public interface ReplicantSessionAuthorization {
     boolean runIfValid(@NonNull Action action) throws IOException;
 
+    /**
+     * Return the Principal associated with the authorized Replicant Session.
+     * Replicant treats the Principal as an opaque application identity.
+     *
+     * @return the Principal.
+     */
     @NonNull
     Object getPrincipal();